Completed the support of colouring the other sprites for multiplayer.
Now, only the colours are missing.

git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@800 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/graphics/CSprite.h
@@ -18,7 +18,7 @@ public:
 	bool optimizeSurface();
 	void applyTransparency();
 	void copy( CSprite &Destination, SDL_Color *Palette );
-	void replaceSpriteColor( Uint32 find, Uint32 replace, Uint16 miny );
+	void replaceSpriteColor( Uint16 find, Uint16 replace, Uint16 miny );
 
 	void setSize(Uint8 w, Uint8 h);
 	void setBouncingBoxCoordinates( Uint16 bboxx1, Uint16 bboxy1, Uint16 bboxx2, Uint16 bboxy2 );
